refactor(recommendations): move api from main package to api package
allow proper import

related to https://github.com/podman-desktop/podman-desktop/issues/14361
