+++ packages/api/src/recommendations/recommendations.ts
@@ -1,5 +1,5 @@
 /**********************************************************************
- * Copyright (C) 2024 Red Hat, Inc.
+ * Copyright (C) 2024-2026 Red Hat, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ***********************************************************************/
 
-import type { FeaturedExtension } from '/@api/featured/featured-api.js';
+import type { FeaturedExtension } from '../featured/featured-api.js';
 
 export interface ExtensionBanner {
   extensionId: string;
