+++ packages/renderer/src/lib/recommendation/ExtensionBanner.spec.ts
@@ -23,8 +23,7 @@ import { tick } from 'svelte';
 import { beforeEach, describe, expect, test, vi } from 'vitest';
 
 import ExtensionBanner from '/@/lib/recommendation/ExtensionBanner.svelte';
-
-import type { ExtensionBanner as IExtensionBanner } from '../../../../main/src/plugin/recommendations/recommendations-api';
+import type { ExtensionBanner as IExtensionBanner } from '/@api/recommendations/recommendations';
 
 const baseBanner: IExtensionBanner = {
   extensionId: 'extension.banner',
