+++ packages/renderer/src/lib/recommendation/ExtensionBanner.svelte
@@ -3,8 +3,7 @@ import { CloseButton } from '@podman-desktop/ui-svelte';
 
 import FeaturedExtension from '/@/lib/featured/FeaturedExtension.svelte';
 import type { MessageBoxReturnValue } from '/@api/dialog';
-
-import { type ExtensionBanner } from '../../../../main/src/plugin/recommendations/recommendations-api';
+import { type ExtensionBanner } from '/@api/recommendations/recommendations';
 
 // Pass in the theme appearance colour of PD to the banner, we do it here so we don't have to do multiple isDark checks when rendering multiple banners.
 interface Props {
