+++ packages/renderer/src/lib/recommendation/ExtensionBanners.svelte
@@ -5,8 +5,7 @@ import ExtensionBanner from '/@/lib/recommendation/ExtensionBanner.svelte';
 import { isDark } from '/@/stores/appearance';
 import { extensionBannerInfos } from '/@/stores/extensionBanners';
 import { providerInfos } from '/@/stores/providers';
-
-import type { ExtensionBanner as ExtensionBannerInfo } from '../../../../main/src/plugin/recommendations/recommendations-api';
+import type { ExtensionBanner as ExtensionBannerInfo } from '/@api/recommendations/recommendations';
 
 let banners: ExtensionBannerInfo[] = $derived.by(() =>
   $extensionBannerInfos.filter(banner => !banner.when || isBannerVisible(banner)),
